A dataset of depth profiles (10Be)

The general characteristics of each sites and studies are only reported for the first item (label=1)

tcn_depth_profiles

Format

A dataframe

study

Study key word

sample

Sample name

label

Sample number

depth

Depth below surface (cm)

C

10Be concentration (at/g)

C_e

Uncertainty on 10Be concentration (at/g)

latitude

Site latitude (°) only reported for first sample (label=1)

altitude

Site altitude (m) only reported for first sample (label=1)

Details

Siame, L., Bellier, O., Braucher, R., Sébrier, M., Cushing, M., Bourlès, D., Hamelin, B., Baroux, E., de Voogd, B., Raisbeck, G., & Yiou, F. (2004). Local erosion rates versus active tectonics: cosmic ray exposure modelling in Provence (south-east France). Earth and Planetary Science Letters, 220(3–4), 345–364. http://dx.doi.org/10.1016/S0012-821X(04)00061-5

Hidy, A. J., Gosse, J. C., Pederson, J. L., Mattern, J. P., & Finkel, R. C. (2010). A geologically constrained Monte Carlo approach to modeling exposure ages from profiles of cosmogenic nuclides: An example from Lees Ferry, Arizona. Geochemistry Geophysics Geosystems, 11, Q0AA10. https://doi.org/10.1029/2010GC003084

Laloy, E., Beerten, K., Vanacker, V., Christl, M., Rogiers, B., & Wouters, L. (2017). Bayesian inversion of a CRN depth profile to infer Quaternary erosion of the northwestern Campine Plateau (NE Belgium). Earth Surface Dynamics, 5(3), 331–345. https://doi.org/10.5194/esurf-5-331-2017